TG-117: change routes and name of superadmin folder
