+++ src/Components/Admin/SuperAdmins/Form/SuperAdminForm.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from 'Components/SuperAdmin/SuperAdmins/Form/superAdminForm.module.css';
+import styles from 'Components/Admin/SuperAdmins/Form/superAdminForm.module.css';
 import Button from 'Components/Shared/Buttons/Buttons';
 import Input from 'Components/Shared/Field/Input';
 import RadioButton from 'Components/Shared/Field/RadioButton';
